+++ src/main/java/com/nyx/bot/modules/warframe/controller/EphemerasController.java
@@ -1,4 +1,4 @@
-package com.nyx.bot.modules.warframe.controller.data;
+package com.nyx.bot.modules.warframe.controller;
 
 import com.fasterxml.jackson.annotation.JsonView;
 import com.nyx.bot.common.core.AjaxResult;
@@ -6,8 +6,8 @@ import com.nyx.bot.common.core.Views;
 import com.nyx.bot.common.core.controller.BaseController;
 import com.nyx.bot.common.core.page.TableDataInfo;
 import com.nyx.bot.data.WarframeDataSource;
-import com.nyx.bot.modules.warframe.entity.RivenItems;
-import com.nyx.bot.modules.warframe.repo.RivenItemsRepository;
+import com.nyx.bot.modules.warframe.entity.Ephemeras;
+import com.nyx.bot.modules.warframe.repo.EphemerasRepository;
 import com.nyx.bot.utils.I18nUtils;
 import jakarta.annotation.Resource;
 import org.springframework.data.domain.PageRequest;
@@ -19,29 +19,25 @@ import org.springframework.web.bind.annotation.RestController;
 import java.util.concurrent.CompletableFuture;
 
 @RestController
-@RequestMapping("/data/warframe/market/riven")
-public class MarketRivenController extends BaseController {
+@RequestMapping("/data/warframe/ephemeras")
+public class EphemerasController extends BaseController {
     @Resource
-    RivenItemsRepository repository;
+    EphemerasRepository ephemerasRepository;
 
     @PostMapping("/list")
     @JsonView(Views.View.class)
-    public TableDataInfo list(@RequestBody RivenItems rivenItems) {
-        return getDataTable(
-                repository.findAllPageable(
-                        rivenItems.getName(),
-                        rivenItems.getRivenType(),
-                        PageRequest.of(
-                                rivenItems.getCurrent() - 1,
-                                rivenItems.getSize())
-                )
-        );
+    public TableDataInfo list(@RequestBody Ephemeras e) {
+        return getDataTable(ephemerasRepository.findAllPageable(
+                e.getName(),
+                PageRequest.of(
+                        e.getCurrent() - 1,
+                        e.getSize())
+        ));
     }
 
     @PostMapping("/update")
     public AjaxResult update() {
-        CompletableFuture.runAsync(WarframeDataSource::getRivenWeapons);
+        CompletableFuture.runAsync(WarframeDataSource::getEphemeras);
         return success(I18nUtils.RequestTaskRun());
     }
-
 }
