+++ src/main/java/com/nyx/bot/modules/warframe/controller/RelicsController.java
@@ -1,4 +1,4 @@
-package com.nyx.bot.modules.warframe.controller.data;
+package com.nyx.bot.modules.warframe.controller;
 
 import com.nyx.bot.common.core.controller.BaseController;
 import com.nyx.bot.common.core.page.TableDataInfo;
